removed errors in cart

+++ src/main/java/com/devops/thatguy/dao/CartDAO.java
@@ -13,7 +13,8 @@ public void delete(int userId);
 public long CartPrice(int userId);	
 public Cart getitem(int prodId,int userId);
 
-public int cartsize(int userId);
+//public int cartsize(int userId);
+public long cartsize(int userId);
 
 public void pay(int userId);
 public List<Cart> AdminReport();
